put in send window updates (snd.wnd, snd.wnd1, snd.wnd2)

+++ headers/contexts/TCPTahoeCongestionControlContext.h
@@ -19,8 +19,8 @@ public:
     TCPTahoeCongestionControlContext(u_int32_t iss);
     virtual ~TCPTahoeCongestionControlContext();
 
-    u_int32_t get_snd_wnd() const;
-    void set_snd_wnd(u_int32_t snd_wnd);
+    u_int16_t get_snd_wnd() const;
+    void set_snd_wnd(u_int16_t snd_wnd);
 
     u_int32_t get_snd_wnd1() const;
     void set_snd_wnd1(u_int32_t snd_wnd1);
@@ -29,7 +29,7 @@ public:
     void set_snd_wnd2(u_int32_t snd_wnd2);
     
 private:
-    u_int32_t snd_wnd_;
+    u_int16_t snd_wnd_;
     u_int32_t snd_wnd1_;
     u_int32_t snd_wnd2_;
